+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Lost Dutchman Mine Change Log
 0.0.3 April 7, 2024
 -------------------
 
-
+- Enh #82: Adjustments to splash screen (MadejaMaciej)
 
 0.0.2 March 24, 2024
 -------------------